requirements.txt
anthropic>=0.23.0
pandas>=1.3.0,<2.0.0
rich>=12.0.0
python-dotenv>=0.19.0
streamlit>=1.27.0,<1.34.0
